Fix the Breaker and remove some tabs
The circuit breker logic was doing something funny where a half closed
state wasn't tripping back closed even if the call was a valid one to
be made.  It would seem that this was a brain fart when I did the recent
version as the functionality DID change when it wasn't intended to.

Remove tabs in Sampled Quantile.