Attempt to address the inaccuracies of the 95th, 99th measures
In practical use - the sampled nature of the performance monitor was making the higher percentiles inaccurate by as much as 10% which isn't really awesome. A lab week project at one point attempted to address that issue by changing the way that the values are sampled - this is the result. Would love to get some people to look at and comment on this idea.